Initial Commit - Completed App - Ready For Review
Created an 11ty project to work as a blog site for Aunt Rosie. This project includes the needed SQL Files to create the table and connection to the Supabase Database we are using for the project. Bootstrap style was also added to deliver the product quicker, can be adjusted. Features a login-logout functionality, as well as creating blog post functionality.